Split out the schema preparation and update logic into its own module
