+++ synapse/storage/engines/sqlite3.py
@@ -13,7 +13,9 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-from synapse.storage import prepare_database, prepare_sqlite3_database
+from synapse.storage._schema_prepare import (
+    prepare_database, prepare_sqlite3_database
+)
 
 
 class Sqlite3Engine(object):
